fixed for
If Late or Absent is preselected - the dropdown automatically opens

Z- I think the rules for the dropdown need to be refined. .5 sec delay, only open if you clicked on it, and when you scroll all the way to the bottom and hit down arrow again it moves to the next kid.